requirements.txt
sentence-transformers
faiss-cpu
python-docx
python-pptx
pdfplumber
openai
langchain
